### 2024-06-14 0c8e566
Allow li method to take dot arguments
