+++ R/fit.R
@@ -86,7 +86,7 @@ skipTrack.fit <- function(Y,cluster,
     #Run skipTrack.MCMC on each worker (or liMCMC if li == TRUE)
     res <- foreach::foreach(1:chains) %dopar% {
       if(li){
-        liMCMC(Y = Y, cluster = cluster, reps = reps, hyperparams = par, S = numSkips)
+        liMCMC(Y = Y, cluster = cluster, reps = reps, hyperparams = par, S = numSkips, ...)
       }else{
         skipTrack.MCMC(Y = Y, cluster = cluster, X = X, Z = Z, numSkips = numSkips, reps = reps,
                       ...)
