### 2024-06-14 0198b33
Allow li method to take dot arguments2
