### 2024-06-25 #6928 @editorialbot
[REVIEW]: skipTrack: An R package for Identifying Skips in Self-Tracked Mobile Menstrual Cycle Data
**Submitting author:** @LukeDuttweiler (<a href="http://orcid.org/0000-0002-0467-995X">Luke Duttweiler</a>)
**Repository:** https://github.com/LukeDuttweiler/skipTrack
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
