### 2024-06-25 #6928 @skanwal
@ankurankan, your review will be checklist based. You will have a separate checklist that you should update when carrying out your review.

The reviewer guidelines are available here: [joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Any questions/concerns please let me know.

✨ Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest ✨
