### 2024-07-01 #6928 @nilabhrardas
Hi @LukeDuttweiler, I have had the chance to review your package and would like to share my thoughts. Firstly, the package functions are well-documented, and the paper is well-written. I was able to execute the functions without any errors.
However, I recommend setting a seed before executing skipTrack.simulate() for reproducibility, i.e., I was unable to reproduce the exact results illustrated in the vignette. Further, in the summary section of the vignette, separating the code and the results into different snippets might make the steps more user-friendly.
Overall, I am happy to recommend the paper for publication given the vignette is revised (as per the aforementioned anomalies) or adequate reasoning is provided for not doing so.
