### 2024-07-01 #6928 @LukeDuttweiler
@nilabhrardas Thank you so much for the helpful and quick review! I really appreciate your efforts.

I have updated the vignette according to your comments. Please let me know if there are any other changes you would suggest. Thank you!
