### 2024-07-31 #6928 @ankurankan
Hi @LukeDuttweiler, sorry for the delay in the review. Overall, I think the package and the manuscript are already in great shape. I have a few minor comments:

1. In my opinion, the readability of the "Package Description" section can be improved. Currenlty, the paragraphs in the section seem somewhat disconnected, maybe you could consider listing the features as a bulleted list. I think it would make it much easier to read.
2. Title Case: "Statement of need" -> "Statement of Need" ?
3. I wasn't able to find a community or contribution guide on GitHub. If it's there, could you please point me to it?
