### 2024-07-31 #6928 @LukeDuttweiler
Hi @ankurankan. Thanks for the review!

I've addressed all three of your comments, please let me know if there is anything else I can improve! There was no contribution guide, that has now been included as CONTRIBUTING.md.

Thank you!!
