### 2024-08-05 #6928 @skanwal
@LukeDuttweiler - can you please check and confirm the `Additional Author Tasks After Review is Complete` in the last message?
