+++ NEWS.md
@@ -1,5 +1,9 @@
 # skipTrack (development version)
 
+# skipTrack 0.1.1
+
+* Minor edits to paper and readme based on JOSS review. No functional package changes.
+
 # skipTrack 0.1.0
 
 # skipTrack 0.0.1
