+++ cran-comments.md
@@ -1,15 +1,8 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+0 errors | 0 warnings | 0 notes
 
-* This is a new release.
+#revdepcheck results
 
-## Notes
+No reverse dependencies.
 
-* There are currently no published references for the main methods in this package. We have included references to alternative methods when appropriate, and will add references once papers have been published/submitted to arXiv. 
-
-## Resubmission
-
-This is a resubmission. In this version I have:
-
-* Removed all 'print' calls that were writing to the console without being able to be suppressed, and replaced them with 'warnings' or 'messages' where appropriate.
