+++ NEWS.md
@@ -1,3 +1,5 @@
+# skipTrack (development version)
+
 # skipTrack 0.1.1
 
 # skipTrack 0.1.1
