+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: skipTrack
 Title: A Bayesian Hierarchical Model that Controls for Non-Adherence in Mobile Menstrual Cycle Tracking
-Version: 0.1.1
+Version: 0.1.1.9000
 Authors@R: 
     person("Luke", "Duttweiler", , "lduttweiler@hsph.harvard.edu", role = c("aut", "cre", 'cph'),
            comment = c(ORCID = "0000-0002-0467-995X"))
