### 2024-08-05 #6928 @LukeDuttweiler
@skanwal All tasks done except archiving the release on Zenodo/figshare/etc. Is the GitHub release enough for that requirement? I've never used either of Zenodo or figshare.

Released version is 0.1.1. Thank you!!
