### 2024-08-07 #6928 @skanwal
Hi @LukeDuttweiler - Having a github repo archived somewhere with a DOI is part of the FAIR principles, so I'd recommend doing it. One of my JOSS team members have shared this presentation: https://zenodo.org/records/5553307#.YV7iY3uxVrk - there is a short summary of how to archive your github repo on zenodo from page 43. Hope this will help.
