### 2024-08-07 5cb691f
Remove experimental version tag so that Zenodo will pick up the next gitHub release with the right version.
