### 2024-08-07 #6928 @LukeDuttweiler
@skanwal GitHub Release v0.1.1 is now archived on Zenodo with DOI: 10.5281/zenodo.13260744

Let me know if you need anything else, thank you!
