+++ bibliography.bib
@@ -10,6 +10,7 @@
   number={1},
   pages={3--11},
   year={2022},
+  doi = {10.1093/jamia/ocab182},
   publisher={Oxford University Press}
 }
 
@@ -20,6 +21,7 @@
   volume={118},
   number={5},
   pages={2245--2250},
+  doi = {10.1542/peds.2015-4480},
   year={2006}
 }
 
@@ -31,6 +33,7 @@
   number={1},
   pages={79},
   year={2020},
+  doi = {10.1038/s41746-020-0269-8},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -42,6 +45,7 @@
   number={1},
   pages={83},
   year={2019},
+  doi = {10.1038/s41746-019-0152-7},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -53,6 +57,7 @@
   number={1},
   pages={e50135},
   year={2024},
+  doi = {10.2196/50135},
   publisher={JMIR Publications Inc., Toronto, Canada}
 }
 
@@ -64,6 +69,7 @@
   number={4},
   pages={545--e1},
   year={2022},
+  doi = {10.1016/j.ajog.2021.09.041},
   publisher={Elsevier}
 }
 
@@ -80,5 +86,6 @@ Algorithms},
     author = {Luke Duttweiler},
     year = {2024},
     note = {R package version 0.2.1},
+    doi = {10.32614/cran.package.genmcmcdiag},
     url = {https://CRAN.R-project.org/package=genMCMCDiag},
   }
