### 2024-08-21 #6928 @LukeDuttweiler
Hi @skanwal, apologies for missing that the first time! The references have been updated with the suggested DOIs (which are all correct). Thank you!
