### 2024-09-01 #6928 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
